changes as follows: grouped (Sensors)Readings Java beans into one package (logreadings) and added some methods to help write/read to files
added fileIO packge and added write / read methods to write and read to txt files.
i have worked on binary write but faced problem with desearliaztion as i think it does not work on android easliy

Any how, read/write in txt is working now and the good thing it simpler for debugging.

the read method read the readings from (fromTime , toTime), and return a hash map with datetime as key and arraylist of objects as value. i tested with Accelerometer and it is wokring
it should work for others but i hope u give it a test
i made a commented usage exmaple of the read method in the MainView (in the button) action.
thanks